Introduce a developer mode option in Makefile (#4364)
this option will enable two addtionals for now:
1. Create symbol even for optimized build
2. Enforce warning as error

while I am here, clean up all warnings so that I can get a clean build in developer mode.

+++ src/operator/tensor/matrix_op-inl.h
@@ -74,7 +74,7 @@ inline bool ReshapeShape(const nnvm::NodeAttrs& attrs,
       dshape_vec.push_back(dshape[i]);
     }
     std::vector<int> tmp;
-    int src_idx = 0;
+    size_t src_idx = 0;
     int inf_idx = -1;
     size_t new_size = dshape.Size();
     if (param_.reverse) {
